+++ src/DI/DefinitionRepository.php
@@ -2,6 +2,9 @@
 
 namespace Georgeff\Kernel\DI;
 
+use Georgeff\Kernel\KernelException;
+use Psr\Container\ContainerInterface;
+
 /**
  * @internal
  */
@@ -11,6 +14,11 @@ final class DefinitionRepository
      * @var array<string, DefinitionInterface>
      */
     private array $definitions = [];
+
+    /**
+     * @var array<string, array{factory: callable(ContainerInterface): mixed, innerId: string}>
+     */
+    private array $decorators = [];
 
     public function add(string $id, callable $factory): DefinitionInterface
     {
@@ -23,11 +31,61 @@ final class DefinitionRepository
     }
 
     /**
+     * @param callable(mixed, ContainerInterface): mixed $decorator
+     */
+    public function decorate(string $id, callable $decorator): void
+    {
+        if (isset($this->decorators[$id])) {
+            throw new KernelException("Cannot apply a decorator to an already decorated definition ID: [{$id}]");
+        }
+
+        $innerServiceId = "_{$id}.inner";
+
+        $factory = function (ContainerInterface $c) use ($innerServiceId, $decorator): mixed {
+            return $decorator($c->get($innerServiceId), $c);
+        };
+
+        $this->decorators[$id] = ['factory' => $factory, 'innerId' => $innerServiceId];
+    }
+
+    private function applyDecorator(string $id): void
+    {
+        if (null === ($inner = $this->get($id))) {
+            throw new KernelException("Cannot decorate a non-existing definition ID: [{$id}]");
+        }
+
+        $decorator = $this->decorators[$id];
+
+        $decorated = $this->add($id, $decorator['factory']);
+
+        if ($inner->isShared()) {
+            $decorated->share();
+        }
+
+        foreach ($inner->getAliases() as $alias) {
+            $decorated->alias($alias);
+        }
+
+        foreach ($inner->getTags() as $tag) {
+            $decorated->tag($tag);
+        }
+
+        $this->add($decorator['innerId'], $inner->getFactory());
+    }
+
+    /**
      * @return array<string, DefinitionInterface>
      */
     public function all(): array
     {
         return $this->definitions;
+    }
+
+    public function applyDecorators(): void
+    {
+        foreach (array_keys($this->decorators) as $id) {
+            $this->applyDecorator($id);
+        }
     }
 
     /**